requirements.txt
docling==2.66.0
docling-core==2.57.0
docling-ibm-models==3.10.3
docling-parse==4.7.2
streamlit==1.52.2
psycopg2-binary==2.9.11
neo4j==6.0.3
ollama==0.6.1
pgvector==0.4.2
pandas==2.3.3
python-dotenv==1.2.1
streamlit-agraph==0.0.45
pyvis==0.3.2
pydantic-settings>=2.0
tenacity>=8.2
